split io::write into multiple files
