+++ src/io/write/write.rs
@@ -1,0 +1,23 @@
+use crate::future::Future;
+use crate::task::{Context, Poll};
+
+use std::io;
+use std::pin::Pin;
+
+use futures_io::AsyncWrite;
+
+#[doc(hidden)]
+#[allow(missing_debug_implementations)]
+pub struct WriteFuture<'a, T: Unpin + ?Sized> {
+    pub(crate) writer: &'a mut T,
+    pub(crate) buf: &'a [u8],
+}
+
+impl<T: AsyncWrite + Unpin + ?Sized> Future for WriteFuture<'_, T> {
+    type Output = io::Result<usize>;
+
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        let buf = self.buf;
+        Pin::new(&mut *self.writer).poll_write(cx, buf)
+    }
+}
